Mk/Uses/qt.mk: Complain about components with an unknown suffix
Normally, USE_QT complains about unknown components with

	===>  $PKGNAME cannot be installed: unknown USE_QT component
	'foobar'.

However this fails if the component has a suffix starting with `_`.
For example

	USE_QT=	buildtools_build,core

or even a simple

	USE_QT=	buildtools_buld

is silently accepted, but they do not do anything.  Only components
with _build and _run suffixes should be accepted.

Amend the regular expression we use for this to be stricter.  It
is currently over eager.  Since we only have two cases to cover
here we can spell them out explictly instead.

PR:		235898
Approved by:	kde (tcberner)
Differential Revision:	https://reviews.freebsd.org/D19267


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493464 35697150-7ecd-e111-bb59-0022644237b5