New port: security/softether5: SoftEther VPN 5 (Developer Edition)
SoftEther VPN ("SoftEther" means "Software Ethernet") is an
open-source cross-platform multi-protocol VPN program, created as an
academic project in the University of Tsukuba.  Its protocol is very
fast and it can be used in very restricted environments, as it's able
to transfer packets over DNS and ICMP.  A NAT-Traversal function is
also available, very useful in case the required ports cannot be
opened on the firewall.  The supported third party protocols are
OpenVPN, L2TP/IPSec and SSTP.

This port is a bleeding-edge developer version of SoftEther, not for
production use.  If you're looking for a solid and stable version,
use security/softether (RTM version) or security/softether-devel
(stable version) instead.  This version is available also on other
other than i386/amd64 processors. Users who want to use SoftEther on
single board computers such as Raspberry Pi, BeagleBoard or other
embedded devices, try this port.

WWW: https://www.softether.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493835 35697150-7ecd-e111-bb59-0022644237b5