german/hunspell: Use gsed during the build
hunspell's build currently attempts to use base sed and \t, most
certainly with the expectation that the latter will expand to a
tab.  With base sed, this is currently not the case - '\t' gets
passed through literally to regex(3), which sees this as an escape
of an ordinary character.  Escapes of ordinary characters will be
disallowed in future versions of regex(3), so switch to textproc/gsed
for now which does the right thing with \t.

PR:		233439
Submitted by:	kevans
Approved by:	office (maintainer timeout, > 2 months)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493948 35697150-7ecd-e111-bb59-0022644237b5