- Resurrect net/pen
This is pen, a load balancer for UDP and TCP based protocols such as
DNS, HTTP or SMTP. It allows several servers to appear as one to the
outside and automatically detects servers that are down and distributes
clients among the available servers. This gives high availability and
scalable performance.

PR:		236048
Submitted by:	Ronald Klop
Pull Request:


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@494407 35697150-7ecd-e111-bb59-0022644237b5