New port: net-im/spectral .. a glossy Matrix client.
This is another Qt-based Matrix client, but it is much more focused
on a shiny desktop experience, rather than the more spartan text-like
look of net-im/quaternion.

The source distribution uses git submodules; add only one of them
to GL_*, because we use a standalone net-im/libqmatrixclient .
The PORTVERSION is related to the most-recently-released version,
but it's not particularly accurate: 464 was released four months ago.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@495058 35697150-7ecd-e111-bb59-0022644237b5