Update to the 20180308 snapshot of GCC 8.3.1.
No longer require a not too old version of GCC to build on powerpc64, but
rely on the system compiler (even if that means we need to be explicitly
conservative when it comes to optimizations). [1]

Sync pkg-descr with lang/gcc7-devel, in particular after r442530 there.

PR:		235975 [1]
Submitted by:	Piotr Kubaj <pkubaj@anongoth.pl> [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@495172 35697150-7ecd-e111-bb59-0022644237b5