Update to 17.12.2 This version supports qt5
SvxLink release 17.12 -- 30 Dec 2017
------------------------------------
The biggest news in this release is the initial revision of the SvxReflector
application. The SvxReflector application is meant to be used as a center point
to link SvxLink nodes together. The new SvxLink ReflectorLogic logic core is
used to connect a SvxLink node to the reflector server. One or more logics can
then be connected to the refelctor using normal logic linking.

In addition a lot of smaller feature additions have been made and also a number
of bugs have been fixed. The full list of changes and more details can be found
below. Information on configuration of new features can be found in the manual
pages.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@495462 35697150-7ecd-e111-bb59-0022644237b5