Update to 2.10.7
The upstream configure script has been updated so this will port will now
correctly link against the ports version of OpenSSL if requested.

The --with-perl option no longer needs an argument, it uses the perl binary
to find the correct paths.

The --with-tcl option takes a TCL_LIBDIR argument to specify exactly which
TCL version to use if multiple are installed.

PR:	236881
Submitted by:	caf@bitchx.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@497369 35697150-7ecd-e111-bb59-0022644237b5