Fix build with clang 8.
Workaround clang 8 compile issues on head.  Note that this requires
changes to the kernel source tree in r345196.

While here, switch to using a date for the version number since there
are no real version numbers in the git repository.  This doesn't use
the 'g' prefix since the older versions that already exist would sort
after it.

PR:	236207
Reviewed by:	swills, jrm
Differential Revision:	https://reviews.freebsd.org/D19602


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@497977 35697150-7ecd-e111-bb59-0022644237b5