Update to 2.2.114
Changelog:
This version adds a skin directive <!--IncludeTemplate ... --> and the variable $SkinTemplateIncludeLevel. The core variable documentation format identifiers were moved to the definition term element to allow CSS ":target" styling, and the header and link text of the vardoc table can now be translated. Input forms have a new HTML5 element "tel", a new attribute "pattern" and two bugs were fixed with the classnames of the new elements and with the identifiers of "select" lists. The documentation was updated.

Sponsored by:	Netzkommune GmbH


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@498265 35697150-7ecd-e111-bb59-0022644237b5