deskutils/spice-gtk: add missing LIB_DEPENDS, USES+=gnome
spice-gtk picks up libva when present with no option to disable it easily,
so make it a requisite for a consistent dependency situation in bare-metal
and jail builds.

The FreeBSD ports rig complains that USE_GNOME requires USES=gnome,
so add that.

PR:		237264
Approved by:	xxjack12xx@gmail.com (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@498873 35697150-7ecd-e111-bb59-0022644237b5