Update Dwarf Fortress to latest upstream release, 0.44.12.
Major bug fixes
 - Fixed crash from corrupted units
 - Made baron elevate to count properly
 - Removed intermittent lag from off-site conceptions and births
   (added seasonal unit offloading)
 - Made squad orders respect successful site conquest (so new members
   don't try to reconquer)

Other bug fixes/tweaks
 - Fixed broken display for multiple adventurer agreements
 - Show proper mission text for messengers instead of No Job
 - Made tantruming dwarves cancel activities
 - Controlled how market links are formed somewhat

While here, take maintainership.

Dwarf Fortress is a metaphor for participating in the FreeBSD project:
everything's lovely, until your legendary craftsdwarf falls into the
river starting a tantrum spiral that ends with the walls covered in
blood and the halls echoing with the cries of the dead.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@502750 35697150-7ecd-e111-bb59-0022644237b5