x11/lightdm: Update to 1.30.0, fix crash and change default config
This update addresses a crash in lightdm, related to the way the
user environment for the lightdm process is setup in src/process.c.

The configuration option in lightdm.conf for locked memory is now
honoured - the default configuration has been updated to disable this,
and pkg-message has been updated to reflect this change.

Changes this release:
  https://github.com/CanonicalLtd/lightdm/releases/tag/1.30.0

PR:		238083
Submitted by:	madpilot
Reported by:	David.Boyd49@twc.com
Differential Revision:	https://reviews.freebsd.org/D20541


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@503608 35697150-7ecd-e111-bb59-0022644237b5