Add p5-Template-Plugin-Gettext 0.7
Template::Plugin::Gettext offers an end-to-end localization and
internationalization solution for the Template Toolkit 2. It consists of a
plugin that offers translation functions inside templates and a string extractor
xgettext-tt2 that extracts translatable strings from templates and writes them
to PO files (or rather a .pot file in PO format). The string extractor
xgettext-tt2 is fully customizable and also usable for other i18n plugins or
frameworks for the Template Toolkit.

WWW: https://metacpan.org/release/Template-Plugin-Gettext


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@504449 35697150-7ecd-e111-bb59-0022644237b5