Uses/python.mk: Clarify language, messaging and usage for <version-spec>
Make it clear that <version-spec> is a declarative not imperative
specification to declare what version of Python a port or software
*supports*, not the versions it should *use*.

The version that gets selected is a function of Python.mk, DEFAULT_VERSIONS
which change over time, and can be overridden or otherwise set by the user.

While I'm here, add a special note about bare USES=python (without a
<version-spec>), which is likely to be deprecated at some point down the
line. In the meantime, describe what the semantics of not specifying a
<version-spec> entails, and encourage minimisation of its use where
appropriate.

Approved by:	koobs (python, maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@505730 35697150-7ecd-e111-bb59-0022644237b5