audio/supercollider: Update to 3.10.2
The previous version, 3.9.3, built fine on all FreeBSD releases but the port
itself was less of use in its state because SuperCollider's own interpreter
("sclang") always crashed when indexing various help files. So users were
able to start the graphical IDE ("scide") but were then on their own to get
any use out of it.

The update to 3.10.2 fixes that problem but it comes with some drawbacks:
SuperCollider has shipped Boost libraries of the 1.66 release and those
don't compile with Clang 8. Using the Boost libraries from the ports tree
instead won't work either, as the 3.10.2 release of SuperCollider isn't
compatible with that version, yet.

Switching to GCC also won't work - it builds fine on all releases but
produces broken binaries that segfaults upon invocation. Thus exclude the
build for FreeBSD releases that comes with Clang 8 for a while and revise
the situation when a new release of SuperCollider is available.

Also pass the maintainership to the submitter because the previous
maintainer seems missing in action and there were already more than three
consecutive maintainer timeouts.

PR:		238186
Submitted by:	Neal Nelson <ports@nicandneal.net>
Reviewed by:	tcberner
Approved by:	maintainer timeout (1+ month)
MFH:		2019Q3
Differential Revision:	https://reviews.freebsd.org/D20854


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@506005 35697150-7ecd-e111-bb59-0022644237b5