- Add games/voadi
Humorous 2D adventure game for PC, Mac, and Linux, a hypothetical
question is taken to the extreme. Our heroine Rachel is challenged
by the brutality of nature, the politics of species, and the struggle
to persevere. Rachel's vegan ethos is useful, as the island is
brimming with animals that can help her, but winning their support
is no easy task. Rachel must use her communication skills and
problem-solving abilities-rather than coercion-to get what she
wants.

WWW: https://www.voadi.com/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@507153 35697150-7ecd-e111-bb59-0022644237b5