Add py-xmldiff 2.3
xmldiff is a library and a command-line utility for making diffs out of XML.
This may seem like something that doesn't need a dedicated utility, but change
detection in hierarchical data is very different from change detection in flat
data. XML type formats are also not only used for computer readable0 data, it is
also often used as a format for hierarchical data that can be rendered into
human readable formats. A traditional diff on such a format would tell you line
by line the differences, but this would not be be readable by a human. xmldiff
provides tools to make human readable diffs in those situations.

WWW: https://github.com/Shoobx/xmldiff


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@508265 35697150-7ecd-e111-bb59-0022644237b5