Add rubygem-dalli 2.7.10
Dalli is a high performance pure Ruby client for accessing memcached servers. It
works with memcached 1.4+ only as it uses the newer binary protocol. It should
be considered a replacement for the memcache-client gem.

The name is a variant of Salvador Dali for his famous painting The Persistence
of Memory.

WWW: https://github.com/petergoldstein/dalli


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@508633 35697150-7ecd-e111-bb59-0022644237b5