Install kernel module sources for drm-current-kmod.
Install kernel module sources and makefiles to
$PREFIX/sys/modules/drm-current-kmod.  This permits kernel builds to
build the DRM modules as part of buildkernel via the LOCAL_MODULES
framework.  These modules are installed to the kernel's directory and
are then preferred to the generic modules installed in /boot/modules.

The installation of sources is controlled by a new SOURCES option
which defaults to on.

Reviewed by:	zeising, manu
Differential Revision:	https://reviews.freebsd.org/D20990


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@508877 35697150-7ecd-e111-bb59-0022644237b5