Unbreak sysutils/pc-networkmanager
This is part of the TrueOS tools, and it was marked BROKEN about
a year ago; a fix appeared upstream only days later, but there's
been no udpate to the port.

The same kind of fix probably applies to sysutils/pcbsd-utils-qt5
but I don't feel like chasing it. PR listed below is for that port,
pinging maintainer(s).

(All this is because we're chasing fallout for the Qt 5.13 update)

PR:		231682


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@510742 35697150-7ecd-e111-bb59-0022644237b5