accessibility/orca: Fix pkg-descr ending with a URL
A URL ending a sentence, with a full stop after it, causes pain
for renderers such as freshports.org and other clients that may hyperlink
them.

Since the same URL is just below it in the pkg-descr WWW URL, amend
the sentence to point to it instead.

PR:		240385
Reported by:	Graham Perrin <grahamperrin gmail com>
Approved by:	portmgr (blanket: just fix it)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@511392 35697150-7ecd-e111-bb59-0022644237b5