New port: devel/termbox Legacy-free alternative to ncurses
Termbox is a library that provides minimalistic API which allows the
programmer to write text-based user interfaces.

It is based on a very simple abstraction. The main idea is viewing
terminals as a table of fixed-size cells and input being a stream of
structured messages. Would be fair to say that the model is inspired
by windows console API. The abstraction itself is not perfect and it
may create problems in certain areas. The most sensitive ones are
copy & pasting and wide characters (mostly Chinese, Japanese, Korean
(CJK) characters). When it comes to copy & pasting, the notion of
cells is not really compatible with the idea of text. And CJK runes
often require more than one cell to display them nicely. Despite the
mentioned flaws, using such a simple model brings benefits in a form
of simplicity. And KISS principle is important.

At this point one should realize, that CLI (command-line interfaces)
aren't really a thing termbox is aimed at. But rather
pseudo-graphical user interfaces.

WWW: https://github.com/nsf/termbox

PR:		240522
Submitted by:	as@php.net
Differential Revision:	https://reviews.freebsd.org/D21614


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@512156 35697150-7ecd-e111-bb59-0022644237b5